+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add benchmarks for merkle opening proof [#197]
 
+### Changed
+
+- Derive `Copy` for `PoseidonBranch` [#200]
+
 ## [0.28.0]
 
 ### Changed
@@ -353,6 +357,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Variants of sponge for `Scalar` & `Gadget(Variable/LC)`.
 
 <!-- ISSUES -->
+[#200]: https://github.com/dusk-network/poseidon252/issues/200
 [#197]: https://github.com/dusk-network/Poseidon252/issues/197
 [#189]: https://github.com/dusk-network/poseidon252/issues/189
 [#181]: https://github.com/dusk-network/poseidon252/issues/181